Make useApproximateExp default to false.
It turns out the approximate exps don't play with the optimizer,
as they cause the likelihood space to be non-smooth at a level the
optimizer can see.  Fixing this is harder than just tweaking the
convergence criteria, so I'm disabling the fast exps by default until
I have time to investigate further.